dig animation & mechanic
