+++ app/core/pokemon-entity.ts
@@ -46,9 +46,6 @@ import AttackingState from "./attacking-state"
 import type { Board } from "./board"
 import {
   Effect as EffectClass,
-  FireHitEffect,
-  GrowGroundEffect,
-  MonsterKillEffect,
   OnAttackEffect,
   OnDamageReceivedEffect,
   OnHitEffect,
@@ -58,6 +55,11 @@ import {
 } from "./effects/effect"
 import { ItemEffects } from "./effects/items"
 import { PassiveEffects } from "./effects/passives"
+import {
+  FireHitEffect,
+  GrowGroundEffect,
+  MonsterKillEffect
+} from "./effects/synergies"
 import { IdleState } from "./idle-state"
 import { ItemStats } from "./items"
 import MovingState from "./moving-state"
@@ -307,33 +309,35 @@ export class PokemonEntity extends Schema implements IPokemonEntity {
         attacker &&
         !attacker.items.has(Item.PROTECTIVE_PADS)
       ) {
-        this.commands.push(new DelayedCommand(() => {
-          const bounceCrit =
-            crit ||
-            (this.effects.has(EffectEnum.ABILITY_CRIT) &&
-              chance(this.critChance, this))
-          const bounceDamage = Math.round(
-            ([0.5, 1][this.stars - 1] ?? 1) *
-            damage *
-            (1 + this.ap / 100) *
-            (bounceCrit ? this.critPower : 1)
-          )
-          this.broadcastAbility({
-            skill: attacker.skill,
-            positionX: this.positionX,
-            positionY: this.positionY,
-            targetX: attacker.positionX,
-            targetY: attacker.positionY
-          })
-          // not handleSpecialDamage to not trigger infinite loop between two magic bounces
-          attacker.handleDamage({
-            damage: bounceDamage,
-            board,
-            attackType: AttackType.SPECIAL,
-            attacker: this,
-            shouldTargetGainMana: true
-          })
-        }, 500))
+        this.commands.push(
+          new DelayedCommand(() => {
+            const bounceCrit =
+              crit ||
+              (this.effects.has(EffectEnum.ABILITY_CRIT) &&
+                chance(this.critChance, this))
+            const bounceDamage = Math.round(
+              ([0.5, 1][this.stars - 1] ?? 1) *
+              damage *
+              (1 + this.ap / 100) *
+              (bounceCrit ? this.critPower : 1)
+            )
+            this.broadcastAbility({
+              skill: attacker.skill,
+              positionX: this.positionX,
+              positionY: this.positionY,
+              targetX: attacker.positionX,
+              targetY: attacker.positionY
+            })
+            // not handleSpecialDamage to not trigger infinite loop between two magic bounces
+            attacker.handleDamage({
+              damage: bounceDamage,
+              board,
+              attackType: AttackType.SPECIAL,
+              attacker: this,
+              shouldTargetGainMana: true
+            })
+          }, 500)
+        )
       }
       return { death: false, takenDamage: 0 }
     } else {
@@ -1755,8 +1759,7 @@ export class PokemonEntity extends Schema implements IPokemonEntity {
     targetX?: number
     targetY?: number
     delay?: number
-  }
-  ) {
+  }) {
     const room = this.simulation.room
     const players = room.state.players
     for (const client of room.clients) {
