+++ app/models/colyseus-models/player.ts
@@ -9,7 +9,7 @@ import {
 import { PokemonEntity } from "../../core/pokemon-entity"
 import type GameState from "../../rooms/states/game-state"
 import { IPlayer, Role, Title } from "../../types"
-import { SynergyTriggers, UniquePool } from "../../types/Config"
+import { BOARD_HEIGHT, BOARD_WIDTH, SynergyTriggers, UniquePool } from "../../types/Config"
 import { DungeonDetails, DungeonPMDO } from "../../types/enum/Dungeon"
 import { BattleResult, Rarity, Team } from "../../types/enum/Game"
 import {
@@ -19,6 +19,7 @@ import {
   Item,
   ItemComponents,
   SynergyGivenByItem,
+  SynergyStones,
   TMs,
   WeatherRocks
 } from "../../types/enum/Item"
@@ -40,8 +41,10 @@ import { getPokemonCustomFromAvatar } from "../../utils/avatar"
 import { getFirstAvailablePositionInBench, isOnBench } from "../../utils/board"
 import { max, min } from "../../utils/number"
 import {
+  chance,
   pickNRandomIn,
   pickRandomIn,
+  shuffleArray,
   simpleHashSeededCoinFlip
 } from "../../utils/random"
 import { resetArraySchema, values } from "../../utils/schemas"
@@ -104,6 +107,7 @@ export default class Player extends Schema implements IPlayer {
     pickRandomIn(Berries)
   ]
   @type(["uint8"]) berryTreesStage: number[] = [1, 1, 1]
+  @type(["uint8"]) groundHoles: number[] = new Array(BOARD_WIDTH * BOARD_HEIGHT / 2).fill(0)
   @type("string") map: DungeonPMDO | "town"
   @type({ set: "string" }) effects: Effects = new Effects()
   @type(["string"]) regionalPokemons = new ArraySchema<Pkm>()
@@ -122,6 +126,7 @@ export default class Player extends Schema implements IPlayer {
   opponents: Map<string, number> = new Map<string, number>()
   titles: Set<Title> = new Set<Title>()
   artificialItems: Item[] = pickNRandomIn(ArtificialItems, 3)
+  buriedItems: (Item | null)[] = initBuriedItems()
   tms: (Item | null)[] = pickRandomTMs()
   weatherRocks: Item[] = []
   randomComponentsGiven: Item[] = []
@@ -606,6 +611,34 @@ function pickRandomTMs() {
   const secondTM = pickRandomIn(TMs.filter((tm) => tm !== firstTM))
   const hm = pickRandomIn(HMs)
   return [firstTM, secondTM, hm]
+}
+
+function initBuriedItems() {
+  /*
+  3 synergy gems
+  3 trash (Trash, Leftovers, Coin)
+  1 precious (artificial item, treasure box, ghimighoul coin)
+  */
+  const buriedItems: (Item | null)[] = new Array(24).fill(null)
+
+  // 3 synergy gems
+  for (let i = 0; i < 3; i++) {
+    buriedItems[i] = pickRandomIn(SynergyStones)
+  }
+
+  // 3 trash (Trash, Leftovers, Coin)
+  for (let i = 3; i < 6; i++) {
+    buriedItems[i] = pickRandomIn([Item.TRASH, Item.LEFTOVERS, Item.COIN, Item.FOSSIL_STONE])
+  }
+
+  // 1 precious (artificial item, treasure box, ghimighoul coin)
+  buriedItems[6] = chance(1 / 2) ? pickRandomIn(ArtificialItems) : pickRandomIn([
+    Item.TREASURE_BOX,
+    Item.GIMMIGHOUL_COIN
+  ])
+
+  shuffleArray(buriedItems)
+  return buriedItems
 }
 
 function spawnDIAYAvatar(player: Player): Pokemon {
