+++ app/public/src/game/depths.ts
@@ -38,7 +38,7 @@ export const DEPTH = {
   // 51-60: Ingame texts/indicators
   DAMAGE_PORTRAIT: 51,
   DAMAGE_TEXT: 52,
-
+  ITEM_FOUND: 53,
   TEXT_MINOR: 54,
   TEXT: 55,
   TEXT_MAJOR: 56,
